move masterdata to single file
