+++ src/app/domains/individual/feature-description/individual-description-header.widget.ts
@@ -4,11 +4,10 @@ import { MatIconButton } from '@angular/material/button';
 import { MatIcon } from '@angular/material/icon';
 import { MatTooltip } from '@angular/material/tooltip';
 import { RouterLink } from '@angular/router';
-import { MasterdataService } from '@masterdata/masterdata.service';
-import { Phenophase } from '@masterdata/phaenophase.model';
-import { Species } from '@masterdata/species.model';
 import { TranslateModule } from '@ngx-translate/core';
 import { Individual } from '@shared/models/individual.model';
+import { Phenophase, Species } from '@shared/models/masterdata.model';
+import { MasterdataService } from '@shared/models/masterdata.service';
 import { PublicUserService } from '@shared/services/public-user.service';
 import { ShortdatePipe } from '@shared/utils/shortdate.pipe';
 import { combineLatest, Observable } from 'rxjs';
