+++ src/app/domains/individual/individual.page.ts
@@ -4,9 +4,9 @@ import { MatDialog } from '@angular/material/dialog';
 import { ActivatedRoute, Router } from '@angular/router';
 import { BaseDetailComponent } from '@core/components/base-detail.component';
 import { AuthService } from '@core/services/auth.service';
-import { MasterdataService } from '@masterdata/masterdata.service';
 import { NavService } from '@shared/components/nav.service';
 import { Individual } from '@shared/models/individual.model';
+import { MasterdataService } from '@shared/models/masterdata.service';
 import { Observable, combineLatest } from 'rxjs';
 import { filter, map } from 'rxjs/operators';
 import { IndividualDescriptionBasicInfoComponent } from './feature-description/basic-info.widget';
