+++ src/app/domains/individual/shared/individual.service.ts
@@ -6,10 +6,8 @@ import { BaseResourceService } from '@core/services/base-resource.service';
 import { FirestoreDebugService } from '@core/services/firestore-debug.service';
 import { IndividualPhenophase } from '@shared/models/individual-phenophase.model';
 import { Individual, SensorLiveData } from '@shared/models/individual.model';
-import { IdLike } from '@shared/models/masterdata/masterdata-like.model';
-import { MasterdataService } from '@shared/models/masterdata/masterdata.service';
-import { Phenophase } from '@shared/models/masterdata/phaenophase.model';
-import { Species } from '@shared/models/masterdata/species.model';
+import { IdLike, Phenophase, Species } from '@shared/models/masterdata.model';
+import { MasterdataService } from '@shared/models/masterdata.service';
 import { AlertService } from '@shared/services/alert.service';
 import { formatShortDate, formatShortDateTime } from '@shared/utils/formatDate';
 import { combineLatest, from, Observable, of } from 'rxjs';
