+++ src/app/domains/individual/station-species-phenophase-observations.model.ts
@@ -1,4 +1,4 @@
-import { Species } from '@masterdata/species.model';
+import { Species } from '@shared/models/masterdata.model';
 import { PhenophaseObservation } from './station-phenophase-observation.model';
 
 export class SpeciesPhenophaseObservations {
