+++ src/app/domains/individual/station.page.ts
@@ -8,13 +8,11 @@ import { MatInput } from '@angular/material/input';
 import { MatSelect, MatSelectChange } from '@angular/material/select';
 import { ActivatedRoute, Router } from '@angular/router';
 import { BaseDetailComponent } from '@core/components/base-detail.component';
-import { Comment } from '@masterdata/comment.model';
-import { MasterdataService } from '@masterdata/masterdata.service';
-import { PhenophaseGroup } from '@masterdata/phaenophase-group.model';
-import { Phenophase } from '@masterdata/phaenophase.model';
 import { TranslateModule } from '@ngx-translate/core';
 import { NavService } from '@shared/components/nav.service';
 import { Individual } from '@shared/models/individual.model';
+import { Comment, Phenophase, PhenophaseGroup } from '@shared/models/masterdata.model';
+import { MasterdataService } from '@shared/models/masterdata.service';
 import { Observation } from '@shared/models/observation.model';
 import { findFirst } from 'fp-ts/lib/Array';
 import _ from 'lodash';
