+++ src/app/domains/map/map.page.ts
@@ -9,11 +9,11 @@ import { MatIcon } from '@angular/material/icon';
 import { MatSelect } from '@angular/material/select';
 import { RouterLink } from '@angular/router';
 import { LocalService } from '@core/services/local.service';
-import { MasterdataService } from '@masterdata/masterdata.service';
-import { Species } from '@masterdata/species.model';
 import { TranslateModule, TranslateService } from '@ngx-translate/core';
 import { NavService } from '@shared/components/nav.service';
 import { MapIndividual } from '@shared/models/individual.model';
+import { Species } from '@shared/models/masterdata.model';
+import { MasterdataService } from '@shared/models/masterdata.service';
 import { SourceFilterType } from '@shared/models/source-type.model';
 import { FormPersistenceService } from '@shared/services/form-persistence.service';
 import { ShortdatePipe } from '@shared/utils/shortdate.pipe';
