+++ src/app/domains/profile/feature-invite/invite-item.widget.ts
@@ -3,8 +3,8 @@ import { Component, Input } from '@angular/core';
 import { MatFabButton } from '@angular/material/button';
 import { MatIcon } from '@angular/material/icon';
 import { RouterLink } from '@angular/router';
-import { MaybeIdLike } from '@masterdata/masterdata-like.model';
 import { TranslateModule } from '@ngx-translate/core';
+import { MaybeIdLike } from '@shared/models/masterdata.model';
 import { ShortdatePipe } from '@shared/utils/shortdate.pipe';
 import { UserSubscriptionButtonComponent } from '../shared/user-subscription.widget';
 import { Invite } from './invite.model';
