+++ src/app/shared/models/individual-phenophase.model.ts
@@ -1,8 +1,6 @@
-import { IdLike } from '@masterdata/masterdata-like.model';
-import { Phenophase } from '@masterdata/phaenophase.model';
-import { Species } from '@masterdata/species.model';
 import { Observable } from 'rxjs';
 import { Individual } from './individual.model';
+import { IdLike, Species, Phenophase } from './masterdata.model';
 
 export class IndividualPhenophase {
   individual: Individual & IdLike;
