--- src/app/shared/models/masterdata/comment.model.ts
@@ -1,7 +1,0 @@
-import { MasterdataLike } from './masterdata-like.model';
-
-export class Comment implements MasterdataLike {
-  id: string;
-  de: string;
-  seq: number;
-}
