--- src/app/shared/models/masterdata/description.model.ts
@@ -1,7 +1,0 @@
-import { MasterdataLike } from './masterdata-like.model';
-
-export class Description implements MasterdataLike {
-  id: string;
-  de: string;
-  seq: number;
-}
