--- src/app/shared/models/masterdata/distance.model.ts
@@ -1,7 +1,0 @@
-import { MasterdataLike } from './masterdata-like.model';
-
-export class Distance implements MasterdataLike {
-  id: string;
-  de: string;
-  seq: number;
-}
