--- src/app/shared/models/masterdata/exposition.model.ts
@@ -1,7 +1,0 @@
-import { MasterdataLike } from './masterdata-like.model';
-
-export class Exposition implements MasterdataLike {
-  id: string;
-  de: string;
-  seq: number;
-}
