--- src/app/shared/models/masterdata/forest.model.ts
@@ -1,7 +1,0 @@
-import { MasterdataLike } from './masterdata-like.model';
-
-export class Forest implements MasterdataLike {
-  id: string;
-  de: string;
-  seq: number;
-}
