--- src/app/shared/models/masterdata/habitat.model.ts
@@ -1,7 +1,0 @@
-import { MasterdataLike } from './masterdata-like.model';
-
-export class Habitat implements MasterdataLike {
-  id: string;
-  de: string;
-  seq: number;
-}
