--- src/app/shared/models/masterdata/irrigation.model.ts
@@ -1,7 +1,0 @@
-import { MasterdataLike } from './masterdata-like.model';
-
-export class Irrigation implements MasterdataLike {
-  id: string;
-  de: string;
-  seq: number;
-}
