--- src/app/shared/models/masterdata/masterdata-like.model.ts
@@ -1,16 +1,0 @@
-export interface MasterdataLike {
-  id: string;
-  de: string;
-}
-
-export interface SortedMasterdataLike extends MasterdataLike {
-  seq: number;
-}
-
-export interface IdLike {
-  id: string;
-}
-
-export interface MaybeIdLike {
-  id?: string;
-}
