--- src/app/shared/models/masterdata/phaenophase-group.model.ts
@@ -1,5 +1,0 @@
-export class PhenophaseGroup {
-  id: string;
-  de: string;
-  seq: number;
-}
