--- src/app/shared/models/masterdata/phaenophase.model.ts
@@ -1,8 +1,0 @@
-export class Phenophase {
-  id: string;
-  description_de: string;
-  de: string;
-  group_id: string;
-  seq: number;
-  comments?: string[];
-}
