--- src/app/shared/models/masterdata/shade.model.ts
@@ -1,7 +1,0 @@
-import { MasterdataLike } from './masterdata-like.model';
-
-export class Shade implements MasterdataLike {
-  id: string;
-  de: string;
-  seq: number;
-}
