--- src/app/shared/models/masterdata/species.model.ts
@@ -1,8 +1,0 @@
-import { SourceType } from '../source-type.model';
-import { MasterdataLike } from './masterdata-like.model';
-
-export class Species implements MasterdataLike {
-  id: string;
-  de: string;
-  sources: SourceType[];
-}
